clean chronometer Callbacks
