+++ clinicadl/callbacks/factory/__init__.py
@@ -1,3 +1,4 @@
+from .chronometer import Chronometer
 from .code_carbon import CodeCarbonCallback
 from .logger import LoggerCallback
 from .mlflow import MLFLOWCallback
